requirements.txt
shapelets-platform
vega-datasets
boto3
altair
